conditions: add `invoker` convenience function
This function creates a simple handler that just invokes the specified
restart (and raises `ControlError` if it cannot be found). The returned
function has the same name as the restart it invokes, to ease debugging.

Unlike Common Lisp, we don't provide `CONTINUE` or `MUFFLE-WARNING` functions;
these can be created (like any custom simple invoker) when needed using
`invoker("proceed")` or `invoker("muffle_warning")`.