Remove Condition base class; conditions are just exceptions.
The condition system already accepts any `Exception` or `Warning` subclass. They
can be signaled, and caught by a `with handlers` form.

The signaling mechanism and Python's exception handling mechanism are separate,
so this causes no confusion on Python's part - signals will fly past exception
handlers and vice versa.

(A technical detail is that the condition system **raises** the internal
`InvokeRestart` exception to transfer control to a restart, but this exception
type is not used for anything else, so the only source of these exceptions is
within the condition system implementation, as is the only sink (provided there
are no catch-all except clauses in the user code). If anyone else chooses to
raise or catch this type of exception, they know what they're doing. In the
other direction, the `error` protocol in the condition system **raises**
`ControlError` to the user as the last resort, if no handler handles the error.
A raise will in any case fly past any `with handlers` forms, so this works as
intended.)

Removing the `Condition` base class is the solution with the most native Python
feel, also avoiding unnecessary class hierarchy duplication. ("What if I want
a ValueErrorCondition?" - just signal a `ValueError` as-is.)

The seemingly reasonable alternative solution of making `Condition` a virtual
base class of `Exception` doesn't work.

If `Condition` inherits from `Exception` and `abc.ABC`, and we use
[`ABCMeta.register`](https://docs.python.org/3/library/abc.html#abc.ABCMeta.register)
to say `Condition.register(Exception)`, this attempts to create an inheritance
cycle, which Python refuses to do.

But if `Condition` inherits only from
[`abc.ABC`](https://docs.python.org/3/library/abc.html#abc.ABC), and does not
have `Exception` as an ancestor, then it gets the wrong default constructor
(from `object`), and is not raiseable.

So we just get rid of `Condition`, and everyone is happy - except purists who
may want to be able to tell apart whether a particular exception class was meant
to be used as a regular exception or as a condition, which is now impossible.
(This is why we originally had the base class.)