Very rough first cut of a REPL server for a live Python process
I.e. the second killer feature of Common Lisp, pythonified.

There are already projects that serve a REPL in a live Python process in various
ways, such as socketserverREPL and Manhole. We need our own mainly to eventually
add support for using syntactic macros in the REPL session connected to the live
process.

(If the user is going to hot-patch code that uses `unpythonic.syntax`, the REPL
must be able to compile code that uses MacroPy macros. Of course it's possible
to replace the relevant source file and just `importlib.reload` the
corresponding module at the REPL, but where's the fun in that?)

We can't use IPython (with imacropy) for this, since
`ipykernel.embed.embed_kernel` will only start from the main thread, which it
then hijacks by entering the Tornado event loop. To work properly in an
arbitrary application, the server should run in a background thread. IPython is
also quite complex to configure.

Since this is mainly an inspection and hot-patching user interface, it doesn't
need to be fancy (i.e. something derived from `code.InteractiveConsole` is
fine), as long as we have basic amenities such as basic line editing, command
history and tab completion.

GNU Readline gives us those, but due to how *nix terminals work (see links in
comments of code added in this commit), and specifically also how Readline works
in Python (see docstring for `rlcompleter`), the `readline` module must run on
the client side.

So we do exactly that, plus (to one-up the simple obvious client solution,
"rlwrap netcat localhost 1337") we provide a remote tab completion facility, via
connecting the local `readline` to a completion backend running on the server.
Our client then talks to both servers, REPL and completion.

Differently from other similar solutions, we run our `InteractiveConsole` inside
a PTY. No idea if that really matters, but at least now it `isatty()`, which
shouldn't hurt. Also, examples beyond `pty.spawn` were not readily available on
the internet, so this also acts as documentation on how to forward traffic
between the master side of a PTY and a network socket in Python (while the slave
side of the PTY is connected to an interactive Python session running in the
same process).

The code works, but is still in a rough state. Mainly, it needs major cleanup
and fixing corner cases regarding connection losses, some more amenities
socketserverREPL's `repl_tool.py` has already figured out (especially prompt
detection to integrate better with `readline`), IPv6 support, and finally, macro
support. There is `macropy.core.MacroConsole`, but to consider this polished,
I want macros to persist within a session, with the same rules `imacropy` uses.
Easiest to do that here first, and then perhaps submit a PR later.

Multiline editing is a question mark, to be investigated. Likely too much
trouble, but won't hurt to see how socketserverREPL does it, if it does it.
Also, our client side needs to understand the difference between a sys.ps1 and a
sys.ps2 prompt (the server already does), so that we can use a blank input on a
new prompt (sys.ps1) as a shortcut for the user to disconnect. (Or maybe that
feature has to go. We'll see.)

This code will also move to a new subpackage, most likely `unpythonic.repl`,
because this prototype suggests we'll need several logical subsystems -
at least REPL server/client, tab completion server/client, PTY/socket proxy.

The implementation of the remote tab completer also suggests a general way to
perform Python RPC via JSON...