Add async_raise: CPython hack to inject exceptions to another thread
Code by Federico Ficarelli and LIU Wei.

We need this for triggering KeyboardInterrupt in a remote REPL session (inside
the session thread, which can be, at that time, stuck waiting in interact()),
when the user presses Ctrl+C at the client side. This and similar awkward
situations in network programming are pretty much the only legitimate use case
for this.