Add ThreadLocalBox and Shim.
ThreadLocalBox is like box, but the contents are thread-local.

Shim holds a box (or a ThreadLocalBox), and redirects attribute accesses on the
shim to whatever object happens to currently be in the box. (E.g. the combo with
ThreadLocalBox can be used to redirect stdin/stdout only in particular threads.)