Fix spurious cache misses in memoizers
We do this by inspecting the parameter bindings `f` would establish if it was
called with the given args and kwargs. The rules are documented in

https://docs.python.org/3/reference/compound_stmts.html#function-definitions
https://docs.python.org/3/reference/expressions.html#calls

so we just implement them independently of Python itself. The result is
`arity.resolve_bindings`.

We also add a dynvar, `resolve_bindings_tuplify`, which can be set to `True`
(via `with dyn.let`) to make `resolve_bindings` return a hashable result. The
default is `False`, in which case it will return a `collections.OrderedDict`
instead, which is nice for end-user use. But memoizers want a hashable result.

Resolves #26.