Use `gensym` where appropriate.
Due to performance reasons, there are still places where it's better to use
`object()`, if the human-readable label is not necessary. One place in
particular that needs to be fast is `arity.resolve_bindings`, because
memoizers must call it every time when handling an incoming call.

Using `gensym` there leads to a significant performance hit in `fix`, noticeable
in the tests (~4-5 seconds instead of 1-2 to run `test_fix.py`, reproducible
not only on CPython, but also on PyPy3).