Get rid of dynvar resolve_bindings_tuplify
A separate `tuplify_bindings` function is a cleaner solution.

We don't need to implicitly pass the flag through several API layers
(which is the use case `dyn` is meant for), and the current return format
of `resolve_bindings` allows performing the tuplification as a separate
postprocessing step - which suggests it should be its own function,
in order to cleanly separate the concerns.