dispatch: check that each method parameter has a type annotation
We should do this if we require type annotations in the multiple-dispatch
system. OTOH, I'm not yet sure if it's a good idea to require them. Dispatching
on just the number of arguments seems useful too, for functions like the builtin
`range` (where the role of a parameter in a particular position depends on how
many arguments were passed in the call).

We should also publish `unpythonic.arity._getfunc`, it's useful whenever there's
a need to look at the signature, and the callable might be not just a regular
function, but an instance method or a class method. Maybe later.