test framework: get rid of evil import loop
It was harmless on Python 3.5+, but 3.4 couldn't figure it out.

Plus this is a better design, with one minor wart: the test macros need to
access the global counters so we count correctly even outside any testset,
but for those the most natural place is in the function layer (to be specific,
in `unpythonic.test.fixtures`).

So the layers can't be fully decoupled. But at least we now observe the
restriction that modules in `unpythonic.syntax`` may import from modules
in `unpythonic`, but not the other way around.