Improve usability of cases like `lazyrec[tuple((1/0, 2/0))]`
Constructors for which the positional mode is "literal_only" are susceptible
to a particular variety of human error.

Without the check in `lazify_ctorcall`, the invocation `lazyrec[tuple(1/0, 2/0)]`
will crash due to a `ZeroDivisionError`. The lazifier skips the arguments, because
the `tuple()` call is wrong; it should be `lazyrec[tuple((1/0, 2/0))]`.

Now we error out the macro expansion when this mistake is detected.