test framework: filter out trivialities from error message
Consider:

test[4 in (1, 2, 3)]       # fails "due to 4 = 4"
test[4 in the[(1, 2, 3)]]  # fails "due to (1, 2, 3) = (1, 2, 3)"

whereas the actual reason is that the expression is False.

So if the captured source code, as a string, matches the repr of the
value (stripping any surrounding whitespace from both), leave out that
triviality when formatting the list of captures for human inspection.

In the above examples, this results in test[] not finding any non-trivial
captures, so it will say the test failed "due to result = False", which
is accurate, and does not confuse the human reader.

This mainly helps new users experimenting with the test framework and trying
pathological examples to get an idea of how to use it.

In real-world usage, interesting subexprs are usually the result of some
computation, or the value of some variable:

x = 4
test[x in (1, 2, 3)]       # fails "due to x = 4"
t = (1, 2, 3)
test[4 in the[t]]          # fails "due to t = (1, 2, 3)"

In such cases, the above issue does not arise.