porting: fix lambdatools bugs
The `quicklambda` macro needs access to the `f` macro interface, to expand that
macro only; so the macro now lives in `unpythonic.syntax.__init__`.

Also, `unpythonic.letdoutil.canonize_bindings` no longer needs a `locref`; this
parameter has been removed. The parameter wasn't very useful in the `mcpyrate`
version, because the source location is not filled in by the time the bindings
are canonized in case the AST being canonized was generated by quasiquotes.