aif's `it` now detects its context at compile time
Also, now that we have `mcpyrate`'s `@namemacro`, we have to be more careful
with parsing name nodes in the AST. It is valid e.g. for a let binding to be
to a name that is actually an expanded name macro (so `mcpyrate` will wrap it
in a `mcpyrate.core.Done` `ASTMarker`).