remove the local `ansicolor` module
It was only used internally by the testing framework, and that framework depends
on `mcpyrate` anyway (since the `test` constructs are macros), so we can just
use `mcpyrate.colorizer`, to DRY out the colorization stuff.

This needs an upgrade to `mcpyrate`, because `colorama` 0.4.4 is missing a
couple of styles, particularly ITALIC. There's new code in `mcpyrate` master
that injects the missing styles (if indeed missing) when the colorizer loads.