Wording for other libraries that provide some feature of unpythonic
The point of having these features in `unpythonic` is integration, and a
consistent API. So if you need only one specific language-extension feature,
then a library that concentrates on that particular feature is likely a good
choice. If you need the kitchen sink, too, then it's better to use our
implementation, since our implementations of the various features are
designed to work together.

In some cases (e.g. the condition system), our implementation may offer extra
features not present in the original library that inspired it.

In other cases (e.g. multiple dispatch), the *other* implementation may be
better (e.g. runs much faster).