Fix wart in continuations subsystem
It used to have to force the return value in `chain_conts`, although in
`unpythonic` return values are never implicitly lazy. This came up in the
Pytkell dialect tests; the dialect itself has `with lazify, autocurry`,
and one test has a `with continuations` block.

By arranging both `lazify` and `autocurry` to leave alone `Values(...)`
as well as `(chain_conts(cc, pcc))(...)`, now we get clean expanded output
also in the continuation-enabled case, and the continuations subsystem sees
enough context to be able to skip lazifying the return value (which is, by
that time, actually an argument to the function returned by `chain_conts`).

Beside now actually honoring the idea that return values are never implicitly
lazy, this also improves performance in code that uses this combination of
features.

Also, improve related comments while at it.