add cautions about Pytkell being slow
Doesn't matter though; this dialect is at its best in surfacing otherwise
hard-to-find bugs or misdesigns in the feature integration of `unpythonic`,
and in `mcpyrate`. I've already fixed a couple of things thanks to the
existence of Pytkell.

Also, it should be serviceable for teaching about lazy functions and autocurry
in a Python-based setting.