In `lazify`, always expand inner macros in recursive mode.
This fixes a bug, for which the test case is:

    from mcpyrate.debug import macros, step_expansion
    from unpythonic.syntax import macros, test

    from unpythonic.syntax import macros, lazify, lazy

    with step_expansion:
        with lazify:
            with test:
                lazy[...]

Without the recursive mode, the `lazy` won't get expanded before `lazify`
performs its own AST edits (editing also `Subscript` nodes, thus breaking macro
invocations), because it's nested inside the `with test` - and `test` expands
outside-in, relying on the expander's recursive mode to expand any remaining
inner macro invocations. (This makes debugging easier, because `step_expansion`
will see it as a step.)

A variant of this test case has been added to `test_lazify.py`.