add unpythonic.syntax._ to make IDEs happy
You can `from unpythonic.syntax import _` to silence any "undefined name" errors
regarding the use of `_`, e.g. in `fn[_ * 3]`.

It's a regular run-time object that does nothing; its only purpose is to give
the name an explicit definition. (Technically, it's an `unpythonic.symbol.sym`,
which seemed appropriate for this.)