markdown: use single backticks
In markdown, it renders the same as double backticks, looks cleaner in the
source, and is easier to type on a Finnish keyboard (where the backtick is
behind shift on a dead key, requiring "shift+tick, space" to get *one* backtick).
It also plays more nicely with smartparens-mode in Emacs.